chore(strat-split): split to strategy
