+++ .dockerignore
@@ -7,3 +7,5 @@ dist
 .flaskenv*
 !.env.project
 !.env.vault
+
+.env*
