+++ secrets/cert.pem
@@ -1,0 +1,24 @@
+-----BEGIN CERTIFICATE-----
+MIIEEjCCAnqgAwIBAgIQZKMwXOzmEM3OPf8GfySSOzANBgkqhkiG9w0BAQsFADBj
+MR4wHAYDVQQKExVta2NlcnQgZGV2ZWxvcG1lbnQgQ0ExHDAaBgNVBAsME01TSVxO
+aWNlQE1TSSAoTmljZSkxIzAhBgNVBAMMGm1rY2VydCBNU0lcTmljZUBNU0kgKE5p
+Y2UpMB4XDTI0MDQyNzA3MzAzNVoXDTI2MDcyNzA3MzAzNVowRzEnMCUGA1UEChMe
+bWtjZXJ0IGRldmVsb3BtZW50IGNlcnRpZmljYXRlMRwwGgYDVQQLDBNNU0lcTmlj
+ZUBNU0kgKE5pY2UpMIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAvgpq
+j6DSAdqKxTDtrjMaPnJpvAqd5CdqlKuzmmE8YDcpmZkengSEdT6fcd4V9R/TJkAX
+AADYdK13sgoy3OUoopulVCTQ6r8TCmvktTZRn++riudxyMC+UBwQZ96rsjYJN4Xr
+9TBIcgKKHZsX9Bj72oGuKEbrg6OClJZ2ayOItvJGaCdtgwk3pXKjN81d6g5TWogf
+lGXYq03YyQF/JVKyiNlFzg4pbvNujpOHOtzbjCvS2xRYjpmKrHhZxB0hoe9z9NSd
+cTJkiXXUiCJ7tiXsd7ucelhi+BMft0fmRCa+vFdYKz/T+NZBsAkxayqnCcO3y1Lx
+fiyi3FvAv2MlFiZn5wIDAQABo14wXDAOBgNVHQ8BAf8EBAMCBaAwEwYDVR0lBAww
+CgYIKwYBBQUHAwEwHwYDVR0jBBgwFoAUug6Wyjzg7M7IEUx26DF+Y9CVzskwFAYD
+VR0RBA0wC4IJbG9jYWxob3N0MA0GCSqGSIb3DQEBCwUAA4IBgQAHeUp46jKsJebL
+ZdSvTR6gH+L/9E8k2D48p7nwSDSIa8bz7epaAzBM3aCMAQglo7LyxctIJT6+ng4d
+W0lsV6Y6E6AQQz78s6EjB3mp5oiidKCpBZZJDhywhpWLQjkGBGS3vVtb1NpUT4z0
+JXRTMuplT+0Y8xwUe2j9qq5sigWjGJ6v9n3ga9skNGHrp8wgcNtsHm8s1g6oU1IP
+YcGi3dzWYT87WUInTMsoL76oi67GAsU+WAgKxAsS0DGQ+IOICJyiOEhDCADvchQf
+jhGayd52+uY3gYBv7P25WR9RlNdG3PsyBNSRPVU3RrReykKDw+iksZmtHGSDdsJ0
+QMPRMzEw1dRf5POaO7A5S0ALa5e8li/E9HlpSXWzooCnoKe98BgyNag2jN6T0sJI
+NMdnSeonn/xstj6vgs7U6/uawfStISyKKMsGBEjWEq3GoLpWxNxb9OeKoMb/pqUr
+PK4odPMkfk1xqbOYBhq8Dks/r28T1iw61MD0ELeXazCEmMokUBQ=
+-----END CERTIFICATE-----
