+++ src/main.ts
@@ -14,16 +14,19 @@ import { ApiConfigService } from './shared/services/api-config.service';
 import { SharedModule } from './shared/shared.module';
 
 import { OAuthException } from '@/exceptions';
-
 declare const module: any;
 
 async function bootstrap() {
   initializeTransactionalContext();
   const whitelist = [
+    'http://localhost:3000',
+    'https://localhost:3000',
     'http://localhost:8000',
+    'https://localhost:8000',
     'https://e-neighbor.netlify.app',
     'https://sandbox.vnpayment.vn/',
   ];
+
   const app = await NestFactory.create<NestExpressApplication>(
     AppModule,
     new ExpressAdapter(),
@@ -39,10 +42,8 @@ async function bootstrap() {
             whitelist.includes(origin) // Checks your whitelist
             // || !!origin.match(/yourdomain\.com$/) // Overall check for your domain
           ) {
-            console.log('allowed cors for:', origin);
             callback(null, true);
           } else {
-            console.log('blocked cors for:', origin);
             callback(new OAuthException('Not allowed by CORS'), false);
           }
         },
